Refactor: Table CSS
