+++ src/features/spellBook/components/table/spellTable.css
