+++ src/styles/table.css
@@ -1,13 +1,12 @@
-:root
-{
+:root {
     --table-header-background-color:#1C1C1C;
     --table-header-color:#ee5622;
 
     --table-row-primary-background-color:#cdcdcd;
     --table-row-secundary-background-color:#b4b4b4;
 }
-.content-table
-{
+
+.content-table {
     width: 100%;
     border-collapse: collapse;
     border-radius: 5px 5px 0px 0px;
@@ -16,8 +15,8 @@
     overflow: hidden;
 }
 
-.content-table thead tr
-{
+/**************** Table Header Row *****************/
+.content-table thead tr {
     background-color: var(--table-header-background-color);
     color: var(--table-header-color);
     
@@ -26,46 +25,39 @@
     text-align: left;
 }
 
+/**************** Table Header and Cell *****************/
 .content-table th,
-.content-table td
-{
+.content-table td {
     padding: 8px 10px;
 }
 
-.content-table tbody tr
-{
+/**************** Table Body Row *****************/
+.content-table tbody tr {
     background-color: var(--table-row-primary-background-color);
     border-bottom: 1px solid #dddddd;
 }
 
-.content-table tbody tr:nth-of-type(even)
-{
+.content-table tbody tr:nth-of-type(even) {
     background-color: var(--table-row-secundary-background-color);
 }
 
-.content-table tbody tr:last-of-type
-{
+.content-table tbody tr:last-of-type {
     border-bottom: 3px solid var(--table-header-background-color);
 }
 
-@media (min-width: 1024px) 
-{
-    .content-table thead tr
-    {
+@media (min-width: 1024px) {
+    .content-table thead tr {
         font-size: 0.85rem;
     }
 }
 
-@media (min-width: 1536px) 
-{
-    .content-table thead tr
-    {
+@media (min-width: 1536px) {
+    .content-table thead tr {
         font-size: 0.9rem;
     }
 
     .content-table th,
-    .content-table td
-    {
+    .content-table td {
         padding: 12px 15px;
     }
 }
