+++ src/styles/table_tabs.css
@@ -1,5 +1,4 @@
-.table-tab
-{
+.table-tab {
     padding: 8px;
     border: 1px solid var(--table-button-border-color);
     border-radius: 4px 4px 0px 0px;
@@ -14,26 +13,21 @@
 }
 
 .table-tab:hover,
-.table-tab:focus
-{
+.table-tab:focus {
     scale: 1.1;
     border-color: var(--table-button-focus-border-color);
     background-color: var(--table-button-focus-background-color);
     color: var(--table-button-focus-color);
 }
 
-@media (min-width: 640px) 
-{
-    .table-tab
-    {
+@media (min-width: 640px) {
+    .table-tab {
         font-size: 0.9rem;
     }
 }
 
-@media (min-width: 1024px) 
-{
-    .table-tab
-    {
+@media (min-width: 1024px) {
+    .table-tab {
         font-size: 1rem;
     }
 }
